backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
google-generativeai
sentence-transformers
PyPDF2
python-docx
python-pptx
pandas
markdown
numpy
pydantic
python-jose
openpyxl
pydantic
python-dotenv
uuid
asyncio
aiofiles